Use prepopulated store for quotes
So, here we are, at the end times, instead of shipping a json file and
the at startup adding all those quotes to the store, the royal we
instead ship a prepopulated store. This wont cut down the size of the
resulting binary by much, or at all, but it will cut down the on device
storage use.